Fixes for Python 3 and installation.
These changes allow Spaun to be run with Python 3 by
* making relative imports explicit
* using function style print
* explicitly converting things to lists where Python 3 returns
  iterators

Furthermore, this fixes some paths so that Spaun can be installed as
a Python package, but this commit does not include a setup.py to do so.

This is the minimal set of changes that were required. Certain
configurations might require more changes.

# Conflicts:
#	_spaun/configurator.py